+++ packages/entities/entities-gateway-services/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.27.9](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-gateway-services%403.27.8...%40kong-ui-public%2Fentities-gateway-services%403.27.9) (2026-09-03)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
 ## [3.27.8](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-gateway-services%403.27.7...%40kong-ui-public%2Fentities-gateway-services%403.27.8) (2026-08-27)
 
 **Note:** Version bump only for package @kong-ui-public/entities-gateway-services