+++ packages/entities/entities-routes/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.26.25](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-routes%403.26.24...%40kong-ui-public%2Fentities-routes%403.26.25) (2026-09-03)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
 ## [3.26.24](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-routes%403.26.23...%40kong-ui-public%2Fentities-routes%403.26.24) (2026-08-27)
 
 **Note:** Version bump only for package @kong-ui-public/entities-routes